fix: update import paths in moved API key components
